pyproject.toml
[project]
name = "speaker-recognition"
version = "0.0.0-dev"
description = "Speaker recognition for Home Assistant using Resemblyzer"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["speaker_recognition*"]

[project.optional-dependencies]
server = [
    "fastapi>=0.68.1",
    "python-dotenv>=0.20.0",
    "torch>=1.9.0,<3.0.0; python_version < '3.10'",
    "resemblyzer>=0.1.3; python_version < '3.10'",
    "setuptools>=59.6.0,<81",
    "typer>=0.10.0",
    "uvicorn>=0.16.0",
]

[dependency-groups]
dev = [
    "mypy>=0.971",
    "ruff>=0.0.17",
    "speaker-recognition[server]",
]
test = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "speaker-recognition[server]",
]

[tool.uv.sources]
torch = { index = "torch" }

[[tool.uv.index]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
filterwarnings = [
  "ignore::DeprecationWarning:resemblyzer.*:",
  "ignore::DeprecationWarning:webrtcvad:",
  "ignore::DeprecationWarning:pkg_resources:",
]
